requirements.txt
altgraph==0.17.4
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
ffmpeg-python==0.2.0
future==1.0.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
m3u8==4.0.0
macholib==1.16.3
moviepy==1.0.3
numpy==1.20.0
packaging==24.0
pillow==10.2.0
proglog==0.1.10
pycryptodome==3.20.0
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
PyQt5==5.15.10
PyQt5-Qt5==5.15.12
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.12
pytube==15.0.0
requests==2.31.0
tqdm==4.66.2
urllib3==2.2.1
